PopulateDB is now SearchDB and we have an import of Google Guava for something called multiMap
