+++ .gitignore
@@ -2,3 +2,4 @@
 /DataAccessLogic/target/
 /DomainModels/target/
 /EnronGUI/target/
+/DataAccessLogic/nbproject/
