+++ BusinessLogic/src/main/java/BusinessLogicLayer/DocumentIndexer.java
@@ -1,0 +1,8 @@
+package BusinessLogicLayer;
+
+import java.nio.file.Path;
+
+public interface DocumentIndexer {
+
+    public void saveDocuments(Path basePath);
+}
