+++ BusinessLogic/src/main/java/BusinessLogicLayer/EnronSearchEngine.java
@@ -1,6 +1,8 @@
 package BusinessLogicLayer;
 
-import DataAccessLayer.Database.PopulateDB;
+import DataAccessLayer.Database.ContainsRepository;
+import DataAccessLayer.Database.DocumentsRepository;
+import DataAccessLayer.Database.TermsRepository;
 import DataAccessLayer.FileSystem.FileLoader;
 import DataAccessLayer.FileSystem.FileLoaderImpl;
 import java.nio.file.Paths;
@@ -22,51 +24,21 @@ public class EnronSearchEngine {
             + FILE_NAME
             + HALF_ALL_DOCS;
 
-    private static FileLoader loader;
-    private static TermSplitter splitter;
-    private static DocumentLoaderImpl documentsLoader;
-    private static PopulateDB populateDB;
-
     public static void main(String[] args) {
-        createDependencies();
+        DocumentIndexer documentsLoader = createDocumentIndexer();
         documentsLoader.saveDocuments(Paths.get(ENRON_DATASET_DIR));
-
-//        List<Document> documents = loadDocuments();
-//        saveDocuments(documents);
     }
 
-    public static void createDependencies() {
-        loader = new FileLoaderImpl();
-        splitter = new TermSplitterImpl("\\s+");
-        populateDB = new PopulateDB();
+    public static DocumentIndexer createDocumentIndexer() {
+        FileLoader loader = new FileLoaderImpl();
+        TermSplitter splitter = new TermSplitterImpl("\\s+");
+        DocumentsRepository documentsRepository = new DocumentsRepository();
+        TermsRepository termsRepository = new TermsRepository();
+        ContainsRepository containsRepository = new ContainsRepository();
 
-        documentsLoader = new DocumentLoaderImpl(loader, splitter, populateDB);
-    }
-
-//    public static List<Document> loadDocuments() {
-//        long startTime = System.nanoTime();
-//
-//
-//        double executionTimeInSeconds = (System.nanoTime() - startTime) / 1E9;
-//
-//        printResult("Load from File System", executionTimeInSeconds,
-//                documentsFromFileSystem.size());
-//
-//        return documentsFromFileSystem;
-//    }
-//    public static void saveDocuments(List<Document> documents) {
-//        long startTime = System.nanoTime();
-//        double executionTimeInSeconds = (System.nanoTime() - startTime) / 1E9;
-//
-//        printResult("Save to DB", executionTimeInSeconds, documents.size());
-//
-//    }
-    private static void printResult(
-            String methodName, double executionTimeInSeconds, int filesSize) {
-        System.out.println(
-                methodName + "!  "
-                + "MethodExecitionTime: " + executionTimeInSeconds
-                + ", Files No: " + filesSize
+        return new DocumentIndexerImpl(
+                loader, splitter,
+                documentsRepository, termsRepository, containsRepository
         );
     }
 }
