+++ pom.xml
@@ -14,4 +14,13 @@
         <module>BusinessLogic</module>
         <module>DomainModels</module>
     </modules>
+    
+    <dependencies>
+        <dependency>
+            <groupId>com.google.guava</groupId>
+            <artifactId>guava</artifactId>
+            <version>19.0</version>
+        </dependency>
+    </dependencies>
+
 </project>
